# build-docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Audiobookshelf API",
    "version": "1.0.0",
    "description": "Audiobookshelf API with autogenerated OpenAPI doc"
  },
  "paths": {
    "/api/authors/{id}": {
      "get": {
        "operationId": "getAuthorByID",
        "summary": "Get a single author by ID on server",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          },
          {
            "$ref": "#/components/parameters/authorInclude"
          },
          {
            "$ref": "#/components/parameters/authorLibraryId"
          }
        ],
        "responses": {
          "200": {
            "description": "getAuthorByID OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/author"
                    },
                    {
                      "$ref": "#/components/schemas/authorWithItems"
                    },
                    {
                      "$ref": "#/components/schemas/authorWithSeries"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      },
      "patch": {
        "operationId": "updateAuthorByID",
        "summary": "Update a single author by ID on server. This endpoint will merge two authors if the new author name matches another author in the database.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          },
          {
            "$ref": "#/components/parameters/asin"
          },
          {
            "$ref": "#/components/parameters/authorName"
          },
          {
            "$ref": "#/components/parameters/authorDescription"
          },
          {
            "$ref": "#/components/parameters/authorImagePath"
          }
        ],
        "responses": {
          "200": {
            "description": "updateAuthorByID OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/author"
                    },
                    {
                      "$ref": "#/components/schemas/authorUpdated"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "merged": {
                          "description": "Will only exist and be `true` if the author was merged with another author",
                          "type": "boolean",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      },
      "delete": {
        "operationId": "deleteAuthorByID",
        "summary": "Delete a single author by ID on server and remove author from all books.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ok200"
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      }
    },
    "/api/authors/{id}/image": {
      "post": {
        "operationId": "setAuthorImageByID",
        "summary": "Set an author image using a provided URL.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          },
          {
            "$ref": "#/components/parameters/imageURL"
          }
        ],
        "responses": {
          "200": {
            "description": "setAuthorImageByID OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/author"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      },
      "delete": {
        "operationId": "deleteAuthorImageByID",
        "summary": "Delete an author image from the server and remove the image from the database.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ok200"
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      },
      "patch": {
        "operationId": "getAuthorImageByID",
        "summary": "Return the author image by author ID.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          },
          {
            "$ref": "#/components/parameters/imageWidth"
          },
          {
            "$ref": "#/components/parameters/imageHeight"
          },
          {
            "$ref": "#/components/parameters/imageFormat"
          },
          {
            "$ref": "#/components/parameters/imageRaw"
          }
        ],
        "responses": {
          "200": {
            "description": "getAuthorImageByID OK",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      }
    },
    "/api/authors/{id}/match": {
      "post": {
        "operationId": "matchAuthorByID",
        "summary": "Match the author against Audible using quick match. Quick match updates the author's description and image (if no image already existed) with information from audible. Either `asin` or `q` must be provided, with `asin` taking priority if both are provided.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/authorID"
          },
          {
            "$ref": "#/components/parameters/asin"
          },
          {
            "$ref": "#/components/parameters/authorSearchName"
          }
        ],
        "responses": {
          "200": {
            "description": "matchAuthorByID OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/author"
                    },
                    {
                      "$ref": "#/components/schemas/authorUpdated"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/author404"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "authorUpdated": {
        "description": "Whether the author was updated without errors. Will not exist if author was merged.",
        "type": "boolean",
        "nullable": true
      },
...
      "authorId": {
        "type": "string",
        "description": "The ID of the author.",
        "format": "uuid",
        "example": "e4bb1afb-4a4f-4dd6-8be0-e615d233185b"
      },
...
      "author": {
        "description": "An author object which includes a description and image path.",
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/authorId"
          },
          "asin": {
            "$ref": "#/components/schemas/authorASIN"
          },
          "name": {
            "$ref": "#/components/schemas/authorName"
          },
          "description": {
            "description": "A description of the author. Will be null if there is none.",
            "type": "string",
            "nullable": true,
            "example": "Terry Goodkind is a #1 New York Times Bestselling Author and creator of the critically acclaimed masterwork,\n‘The Sword of Truth’. He has written 30+ major, bestselling novels, has been published in more than 20\nlanguages world-wide, and has sold more than 26 Million books. ‘The Sword of Truth’ is a revered literary\ntour de force, comprised of 17 volumes, borne from over 25 years of dedicated writing. Terry Goodkind's\nbrilliant books are character-driven stories, with a focus on the complexity of the human psyche. Goodkind\nhas an uncanny grasp for crafting compelling stories about people like you and me, trapped in terrifying\nsituations.\n"
          },
          "imagePath": {
            "description": "The absolute path for the author image located in the `metadata/` directory. Will be null if there is no image.",
            "type": "string",
            "nullable": true,
            "example": "/metadata/authors/aut_bxxbyjiptmgb56yzoz.jpg"
          },
          "addedAt": {
            "$ref": "#/components/schemas/addedAt"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/updatedAt"
          }
        }
      },
      "authorWithItems": {
        "type": "object",
        "description": "The author schema with an array of items they are associated with.",
        "allOf": [
          {
            "$ref": "#/components/schemas/author"
          },
          {
            "type": "object",
            "properties": {
              "libraryItems": {
                "description": "The items associated with the author",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/libraryItemMinified"
                }
              }
            }
          }
        ]
      },
      "authorWithSeries": {
        "type": "object",
        "description": "The author schema with an array of items and series they are associated with.",
        "allOf": [
          {
            "$ref": "#/components/schemas/authorWithItems"
          },
          {
            "type": "object",
            "properties": {
              "series": {
                "description": "The series associated with the author",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/authorSeries"
                }
              }
            }
          }
        ]
      },
...
    },
    "parameters": {
      "authorID": {
        "name": "id",
        "in": "path",
        "description": "Author ID",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/authorId"
        }
      },
      "authorInclude": {
        "name": "include",
        "in": "query",
        "description": "A comma separated list of what to include with the author. The options are `items` and `series`. `series` will only have an effect if `items` is included.",
        "required": false,
        "schema": {
          "type": "string",
          "example": "items"
        },
        "examples": {
          "empty": {
            "summary": "Do not return library items",
            "value": ""
          },
          "itemOnly": {
            "summary": "Only return library items",
            "value": "items"
          },
          "itemsAndSeries": {
            "summary": "Return library items and series",
            "value": "items,series"
          }
        }
      },
      "authorLibraryId": {
        "name": "library",
        "in": "query",
        "description": "The ID of the library to to include filter included items from.",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/libraryId"
        }
      },
      "asin": {
        "name": "asin",
        "in": "query",
        "description": "The Audible Identifier (ASIN).",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/authorASIN"
        }
      },
...
      "authorName": {
        "name": "name",
        "in": "query",
        "description": "The new name of the author.",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/authorName"
        }
      },
      "authorDescription": {
        "name": "description",
        "in": "query",
        "description": "The new description of the author.",
        "required": false,
        "schema": {
          "type": "string",
          "nullable": true,
          "example": "Terry Goodkind is a"
        }
      },
      "authorImagePath": {
        "name": "imagePath",
        "in": "query",
        "description": "The new absolute path for the author image.",
        "required": false,
        "schema": {
          "type": "string",
          "nullable": true,
          "example": "/metadata/authors/aut_z3leimgybl7uf3y4ab.jpg"
        }
      },
      "imageURL": {
        "name": "url",
        "in": "query",
        "description": "The URL of the image to add to the server",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uri",
          "example": "https://images-na.ssl-images-amazon.com/images/I/51NoQTm33OL.__01_SX120_CR0,0,120,120__.jpg"
        }
      },
      "imageWidth": {
        "name": "width",
        "in": "query",
        "description": "The requested width of image in pixels.",
        "schema": {
          "type": "integer",
          "default": 400,
          "example": 400
        },
        "example": 400
      },
...
      "imageFormat": {
        "name": "format",
        "in": "query",
        "description": "The requested output format.",
        "schema": {
          "type": "string",
          "default": "jpeg",
          "example": "webp"
        }
      },
...
    },
    "responses": {
      "ok200": {
        "description": "OK"
      },
      "author404": {
        "description": "Author not found.",
        "content": {
          "text/html": {
            "schema": {
              "type": "string",
              "example": "Not found"
            }
          }
        }
      }
    }
  },
...
}